pyproject.toml
[project]
name = "kadz-discord-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Hung Do",email = "hungdojan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)"
]

[tool.poetry]
packages = [{include = "kadz_discord_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
kadz-discord-bot = "kadz_discord_bot.__main__:main"
